backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# AI/LLM
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
tiktoken==0.5.2

# Vector DB
pinecone-client==3.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
orjson==3.9.12
structlog==24.1.0

# Security
cryptography==42.0.0
PyJWT==2.8.0

# File Handling
Pillow==10.2.0
PyPDF2==3.0.1
python-magic==0.4.27
boto3==1.34.0

# Validation
phonenumbers==8.13.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Celery
celery==5.3.6
flower==2.0.1

# WebSockets
websockets==12.0
